finish

+++ controllers/games.js
@@ -14,14 +14,14 @@ const getAllStaticsGames = async (req, res) => {
 };
 
 const getAllGames = async (req, res) => {
-  const { genre, platform, date, page, limit } = req.query;
+  const { genre, platform, date, page, limit, franchise } = req.query;
   const options = {};
   const select = [
     "title",
     "genre",
     "thumbnail",
     "short_description",
-    "platform",
+    "platform"
   ];
 
   if (genre) {
@@ -36,6 +36,10 @@ const getAllGames = async (req, res) => {
       $gte: `${date}-1-1`,
       $lt: `${parseFloat(parseInt(date) + 1)}-1-1`,
     };
+  }
+
+  if (franchise) {
+    options.franchise = franchise
   }
 
   try {
@@ -70,48 +74,65 @@ const getSingleGame = async (req, res) => {
 const createGame = async (req, res) => {
   const newGame = new BasicGame(req.body);
 
-  try {
-    const savedGame = await newGame.save();
-    res.status(201).json({ status: "created", data: savedGame });
-  } catch (error) {
-    res.status(500).json({
-      status: "create error",
-      message: "ha ocurrido un error al crear el juego",
-      error,
-    });
+  if (req.query.password !== process.env.CONTRA) {
+    res.status(403).json({ status: "error", message: "Invalid Password" });
+  } else {
+    try {
+      const savedGame = await newGame.save();
+      res.status(201).json({ status: "created", data: savedGame });
+    } catch (error) {
+      res.status(500).json({
+        status: "create error",
+        message: "ha ocurrido un error al crear el juego",
+        error,
+      });
+    }
   }
+
 };
 
 const deleteGame = async (req, res) => {
   const { id } = req.params;
-  try {
-    const deletedGame = await BasicGame.findByIdAndDelete(id);
-    if (!deletedGame)
-      return res
-        .status(404)
-        .json({ status: "success", message: "Game not found" });
 
-    res.status(203).json({ status: "success", message: "game eliminated" });
-  } catch (error) {
-    res
-      .status(500)
-      .json({ status: "error", message: "error game not eliminated", error });
+  if (req.query.password !== process.env.CONTRA) {
+    res.status(403).json({ status: "error", message: "Invalid Password" });
+  } else {
+    try {
+      const deletedGame = await BasicGame.findByIdAndDelete(id);
+      if (!deletedGame)
+        return res
+          .status(404)
+          .json({ status: "success", message: "Game not found" });
+  
+      res.status(203).json({ status: "success", message: "game eliminated" });
+    } catch (error) {
+      res
+        .status(500)
+        .json({ status: "error", message: "error game not eliminated", error });
+    }
   }
+
 };
 
 const updateGame = async (req, res) => {
   const { id } = req.params;
   const update = req.body;
-  try {
-    const updatedGame = await BasicGame.findByIdAndUpdate(id, update, {
-      new: true,
-    });
-    res.status(203).json({ status: "success", data: updatedGame });
-  } catch (error) {
-    res
-      .status(500)
-      .json({ status: "error", message: "error game not updated", error });
+
+  if (req.query.password !== process.env.CONTRA) {
+    res.status(403).json({ status: "error", message: "Invalid Password" });
+  } else {
+    try {
+      const updatedGame = await BasicGame.findByIdAndUpdate(id, update, {
+        new: true,
+      });
+      res.status(203).json({ status: "success", data: updatedGame });
+    } catch (error) {
+      res
+        .status(500)
+        .json({ status: "error", message: "error game not updated", error });
+    }
   }
+
 };
 
 module.exports = {
@@ -120,5 +141,5 @@ module.exports = {
   createGame,
   deleteGame,
   updateGame,
-  getAllStaticsGames
+  getAllStaticsGames,
 };
